[FIX] BUG VIEW-1335 Next 15 videos can not be loaded;

+++ ant-viewver/src/main/java/com/antourage/weaverlib/screens/vod/VodPlayerFragment.kt
@@ -118,7 +118,7 @@ internal class VodPlayerFragment : ChatFragment<VideoViewModel>(),
             }
     }
 
-    private val viewersChangeObserver: Observer<Pair<Int, Int>> = Observer { viewInfo ->
+    private val viewersChangeObserver: Observer<Pair<Int, Long>> = Observer { viewInfo ->
         viewInfo?.apply {
             if (first == viewModel.streamId) {
                 txtNumberOfViewers.text = second.toString()
